implemented OpenlayersOsmLayersPlugin and deleted unnecessary functions
