+++ src/app/@ansyn/imagery/model/imap.ts
@@ -10,10 +10,6 @@ export interface IMapConstructor {
 
 	new(...args): IMap;
 
-	addGroupLayer(layer: any, groupName: string);
-
-	removeGroupLayer(layer: any, groupName: string);
-
 	addGroupVectorLayer(layer: any, groupName: string);
 }
 
@@ -27,20 +23,12 @@ export abstract class IMap<T = any> {
 	mapObject: T;
 	projectionService: ProjectionService;
 
-	static addGroupLayer(layer: any, groupName: string) {
-	}
-
-	static removeGroupLayer(id: string, groupName: string) {
-	}
-
 	static addGroupVectorLayer(layer: any, groupName: string) {
 	}
 
 	abstract getCenter(): Observable<Point>;
 
 	abstract setCenter(center: Point, animation: boolean): Observable<boolean>;
-
-	abstract toggleGroup(groupName: string);
 
 	abstract initMap(element: HTMLElement, layers?: any, position?: CaseMapPosition): Observable<boolean>;
 
@@ -70,6 +58,4 @@ export abstract class IMap<T = any> {
 	abstract addGeojsonLayer(data: GeoJsonObject);
 
 	abstract dispose(): void;
-
-	abstract addLayerIfNotExist(layer: any);
 }
