+++ src/app/@ansyn/menu-items/layers-manager/models/layers.model.ts
@@ -6,8 +6,14 @@ export enum LayerType {
 	complex = 'Complex'
 }
 
+export enum layerPluginType {
+	OSM = 'OSM',
+	geoJson = 'geoJson'
+}
+
 export interface ILayer extends Entity {
 	url: string;
 	name: string;
 	type: LayerType;
+	layerPluginType: layerPluginType;
 }
